[fix][broker] Fix potential OOM due to client use sasl authentication

+++ pulsar-broker-common/src/main/java/org/apache/pulsar/broker/ServiceConfigurationUtils.java
@@ -140,7 +140,7 @@ public class ServiceConfigurationUtils {
 
     public static long getLongPropertyOrDefault(ServiceConfiguration config, String key, long defaultValue) {
         Object value = config.getProperties().get(key);
-        if (value instanceof Integer) {
+        if (value instanceof Long) {
             return (Long) value;
         } else if (value instanceof String) {
             try {
